Color Pallet: Move theme pallet screen composable to common ui
